Cardano.Ledger.Allegra.Translation

Documentation

shelleyToAllegraAVVMsToDelete :: NewEpochState (ShelleyEra c) -> UTxO (ShelleyEra c)

Return the subset of UTxO corresponding to Byron-era AVVM addresses, which are to be removed on the Shelley/Allegra boundary. This set will be passed _back_ to the translation functions as the UTxO, allowing these addresses to be removed. This is needed because we cannot do a full scan on the UTxO at this point, since it has been persisted to disk.

Orphan instances

Crypto c => TranslateEra (AllegraEra c) Update

Instance details

Associated Types

type TranslationError (AllegraEra c) Update

Methods

translateEra :: TranslationContext (AllegraEra c) -> Update (PreviousEra (AllegraEra c)) -> Except (TranslationError (AllegraEra c) Update) (Update (AllegraEra c))

Crypto c => TranslateEra (AllegraEra c) WitnessSet

Instance details

Associated Types

type TranslationError (AllegraEra c) WitnessSet

Methods

translateEra :: TranslationContext (AllegraEra c) -> WitnessSet (PreviousEra (AllegraEra c)) -> Except (TranslationError (AllegraEra c) WitnessSet) (WitnessSet (AllegraEra c))

Crypto c => TranslateEra (AllegraEra c) EpochState

Instance details

Associated Types

type TranslationError (AllegraEra c) EpochState

Methods

translateEra :: TranslationContext (AllegraEra c) -> EpochState (PreviousEra (AllegraEra c)) -> Except (TranslationError (AllegraEra c) EpochState) (EpochState (AllegraEra c))

Crypto c => TranslateEra (AllegraEra c) LedgerState

Instance details

Associated Types

type TranslationError (AllegraEra c) LedgerState

Methods

translateEra :: TranslationContext (AllegraEra c) -> LedgerState (PreviousEra (AllegraEra c)) -> Except (TranslationError (AllegraEra c) LedgerState) (LedgerState (AllegraEra c))

Crypto c => TranslateEra (AllegraEra c) UTxOState

Instance details

Associated Types

type TranslationError (AllegraEra c) UTxOState

Methods

translateEra :: TranslationContext (AllegraEra c) -> UTxOState (PreviousEra (AllegraEra c)) -> Except (TranslationError (AllegraEra c) UTxOState) (UTxOState (AllegraEra c))

Crypto c => TranslateEra (AllegraEra c) UTxO

Instance details

Associated Types

type TranslationError (AllegraEra c) UTxO

Methods

translateEra :: TranslationContext (AllegraEra c) -> UTxO (PreviousEra (AllegraEra c)) -> Except (TranslationError (AllegraEra c) UTxO) (UTxO (AllegraEra c))

Crypto c => TranslateEra (AllegraEra c) TxOut

Instance details

Associated Types

type TranslationError (AllegraEra c) TxOut

Methods

translateEra :: TranslationContext (AllegraEra c) -> TxOut (PreviousEra (AllegraEra c)) -> Except (TranslationError (AllegraEra c) TxOut) (TxOut (AllegraEra c))

Crypto c => TranslateEra (AllegraEra c) PPUPState

Instance details

Associated Types

type TranslationError (AllegraEra c) PPUPState

Methods

translateEra :: TranslationContext (AllegraEra c) -> PPUPState (PreviousEra (AllegraEra c)) -> Except (TranslationError (AllegraEra c) PPUPState) (PPUPState (AllegraEra c))

Crypto c => TranslateEra (AllegraEra c) ProposedPPUpdates

Instance details

Associated Types

type TranslationError (AllegraEra c) ProposedPPUpdates

Methods

translateEra :: TranslationContext (AllegraEra c) -> ProposedPPUpdates (PreviousEra (AllegraEra c)) -> Except (TranslationError (AllegraEra c) ProposedPPUpdates) (ProposedPPUpdates (AllegraEra c))

Crypto c => TranslateEra (AllegraEra c) ShelleyGenesis

Instance details

Associated Types

type TranslationError (AllegraEra c) ShelleyGenesis

Methods

translateEra :: TranslationContext (AllegraEra c) -> ShelleyGenesis (PreviousEra (AllegraEra c)) -> Except (TranslationError (AllegraEra c) ShelleyGenesis) (ShelleyGenesis (AllegraEra c))

Crypto c => TranslateEra (AllegraEra c) Tx

Instance details

Associated Types

type TranslationError (AllegraEra c) Tx

Methods

translateEra :: TranslationContext (AllegraEra c) -> Tx (PreviousEra (AllegraEra c)) -> Except (TranslationError (AllegraEra c) Tx) (Tx (AllegraEra c))

Crypto c => TranslateEra (AllegraEra c) NewEpochState

Instance details

Associated Types

type TranslationError (AllegraEra c) NewEpochState

Methods

translateEra :: TranslationContext (AllegraEra c) -> NewEpochState (PreviousEra (AllegraEra c)) -> Except (TranslationError (AllegraEra c) NewEpochState) (NewEpochState (AllegraEra c))

Crypto c => TranslateEra (AllegraEra c) (PParams' f)

Instance details

Associated Types

type TranslationError (AllegraEra c) (PParams' f)

Methods

translateEra :: TranslationContext (AllegraEra c) -> PParams' f (PreviousEra (AllegraEra c)) -> Except (TranslationError (AllegraEra c) (PParams' f)) (PParams' f (AllegraEra c))