Cardano.Ledger.Mary.Value

Documentation

newtype PolicyID crypto

Policy ID

Constructors

PolicyID

Fields

policyID :: ScriptHash crypto

newtype AssetName

Asset Name

Constructors

AssetName

Fields

assetName :: ShortByteString

data Value crypto

The Value representing MultiAssets

Constructors

Value !Integer !(Map (PolicyID crypto) (Map AssetName Integer))

insert :: (Integer -> Integer -> Integer) -> PolicyID crypto -> AssetName -> Integer -> Value crypto -> Value crypto

insert comb policy asset n v, if comb = old new -> old, the integer in the Value is prefered over n if comb = old new -> new, then n is prefered over the integer in the Value if (comb old new) == 0, then that value should not be stored in the Map part of the Value.

lookup :: PolicyID crypto -> AssetName -> Value crypto -> Integer

policies :: Value crypto -> Set (PolicyID crypto)

Extract the set of policies in the Value.

This function is equivalent to computing the support of the value in the spec.

prune :: Map (PolicyID crypto) (Map AssetName Integer) -> Map (PolicyID crypto) (Map AssetName Integer)

Remove 0 assets from a map

representationSize :: forall crypto. Crypto crypto => [(PolicyID crypto, AssetName, Integer)] -> Int

showValue :: Value crypto -> String

Display a Value as a String, one token per line

valueFromList :: Integer -> [(PolicyID era, AssetName, Integer)] -> Value era

Rather than using prune to remove 0 assets, when can avoid adding them in the first place by using valueFromList to construct a Value.

gettriples' :: Value crypto -> (Integer, [(PolicyID crypto, AssetName, Integer)], [PolicyID crypto])

Turn the nested Value map-of-maps representation into a flat sequence of policyID, asset name and quantity, plus separately the ada quantity.