Cardano.Ledger.ShelleyMA.Rules.Utxow

Documentation

class GetPolicies a era where

We want to reuse the same rules for Mary and Allegra. This however relies on being able to get a set of PolicyIDs from the value. Since a Coin has no policies, we create a small class which returns a null set of PolicyIDs for Coin.

This should not escape this module.

Methods

getPolicies ∷ a → Set (PolicyID era)

scriptsNeeded ∷ (ShelleyBased era, GetPolicies (Value era) era, HasField "certs" (TxBody era) (StrictSeq (DCert era)), HasField "wdrls" (TxBody era) (Wdrl era), HasField "inputs" (TxBody era) (Set (TxIn era)), HasField "forge" (TxBody era) (Value era)) ⇒ UTxO era → Tx era → Set (ScriptHash era)

Computes the set of script hashes required to unlock the transaction inputs and the withdrawals.

Orphan instances

(Crypto c, Typeable ma, STS (UTXO (ShelleyMAEra ma c)), BaseM (UTXO (ShelleyMAEra ma c)) ~ ShelleyBase, Compactible (Value (ShelleyMAEra ma c)), Val (Value (ShelleyMAEra ma c)), GetPolicies (Value (ShelleyMAEra ma c)) (ShelleyMAEra ma c), ChainData (Value (ShelleyMAEra ma c)), ChainData (Delta (Value (ShelleyMAEra ma c))), SerialisableData (Value (ShelleyMAEra ma c)), SerialisableData (Delta (Value (ShelleyMAEra ma c))), SerialisableData (CompactForm (Value (ShelleyMAEra ma c))), Torsor (Value (ShelleyMAEra ma c)), DSignable c (Hash c EraIndependentTxBody)) ⇒ STS (UTXOW (ShelleyMAEra ma c))

Instance details

Associated Types

type State (UTXOW (ShelleyMAEra ma c))

type Signal (UTXOW (ShelleyMAEra ma c))

type Environment (UTXOW (ShelleyMAEra ma c))

type BaseM (UTXOW (ShelleyMAEra ma c)) ∷ Type → Type

type PredicateFailure (UTXOW (ShelleyMAEra ma c)) = (b ∷ Type)

Methods

initialRules ∷ [InitialRule (UTXOW (ShelleyMAEra ma c))]

transitionRules ∷ [TransitionRule (UTXOW (ShelleyMAEra ma c))]

assertions ∷ [Assertion (UTXOW (ShelleyMAEra ma c))]

renderAssertionViolation ∷ AssertionViolation (UTXOW (ShelleyMAEra ma c)) → String

(Crypto c, STS (UTXO (ShelleyMAEra ma c)), BaseM (UTXO (ShelleyMAEra ma c)) ~ ShelleyBase) ⇒ Embed (UTXO (ShelleyMAEra ma c)) (UTXOW (ShelleyMAEra ma c))

Instance details

Methods

wrapFailed ∷ PredicateFailure (UTXO (ShelleyMAEra ma c)) → PredicateFailure (UTXOW (ShelleyMAEra ma c))