Cardano.Ledger.ShelleyMA.TxBody

Documentation

newtype TxBody e

Constructors

TxBodyConstr (MemoBytes (TxBodyRaw e))

Bundled Patterns

pattern TxBody :: FamsTo era => Set (TxIn (Crypto era)) -> StrictSeq (TxOut era) -> StrictSeq (DCert (Crypto era)) -> Wdrl (Crypto era) -> Coin -> ValidityInterval -> StrictMaybe (Update era) -> StrictMaybe (AuxiliaryDataHash (Crypto era)) -> Value era -> TxBody era

pattern TxBody' :: Set (TxIn (Crypto era)) -> StrictSeq (TxOut era) -> StrictSeq (DCert (Crypto era)) -> Wdrl (Crypto era) -> Coin -> ValidityInterval -> StrictMaybe (Update era) -> StrictMaybe (AuxiliaryDataHash (Crypto era)) -> Value era -> TxBody era
This pattern is for deconstruction only but accompanied with fields and projection functions.

data TxBodyRaw era

Constructors

TxBodyRaw

Fields

inputs :: !(Set (TxIn (Crypto era)))

outputs :: !(StrictSeq (TxOut era))

certs :: !(StrictSeq (DCert (Crypto era)))

wdrls :: !(Wdrl (Crypto era))

txfee :: !Coin

vldt :: !ValidityInterval

update :: !(StrictMaybe (Update era))

adHash :: !(StrictMaybe (AuxiliaryDataHash (Crypto era)))

mint :: !(Value era)

type FamsFrom era = (Era era, Typeable era, Typeable (Script era), Typeable (AuxiliaryData era), Show (Value era), DecodeNonNegative (Value era), DecodeMint (Value era), FromCBOR (PParams era), FromCBOR (PParamsDelta era), FromCBOR (Value era), FromCBOR (Annotator (Script era)))

type FamsTo era = (Era era, ToCBOR (Value era), EncodeMint (Value era), ToCBOR (Script era), ToCBOR (PParams era), ToCBOR (PParamsDelta era), Typeable (AuxiliaryData era))

txSparse :: FamsTo era => TxBodyRaw era -> Encode ('Closed 'Sparse) (TxBodyRaw era)

bodyFields :: FamsFrom era => Word -> Field (TxBodyRaw era)

data StrictMaybe a

Constructors

SNothing

SJust !a

fromSJust :: StrictMaybe a -> a

data ValidityInterval

ValidityInterval is a half open interval. Closed on the bottom, Open on the top. A SNothing on the bottom is negative infinity, and a SNothing on the top is positive infinity

Constructors

ValidityInterval

Fields

invalidBefore :: !(StrictMaybe SlotNo)

invalidHereafter :: !(StrictMaybe SlotNo)

initial :: Val (Value era) => TxBodyRaw era