Cardano.Ledger.ShelleyMA.Value

Documentation

newtype PolicyID era

Policy ID

Constructors

PolicyID

Fields

policyID ∷ ScriptHash era

newtype AssetID

Asset ID

Constructors

AssetID

Fields

assetID ∷ ByteString

data Value era

The Value representing MultiAssets

Constructors

Value !Integer !(Map (PolicyID era) (Map AssetID Integer))

insert ∷ (Integer → Integer → Integer) → PolicyID era → AssetID → Integer → Value era → Value era

insert comb policy asset n v, if comb = old new -> old, the integer in the Value is prefered over n if comb = old new -> new, then n is prefered over the integer in the Value if (comb old new) == 0, then that value should not be stored in the Map part of the Value.

lookup ∷ PolicyID era → AssetID → Value era → Integer

showValue ∷ Value era → String

Display a Value as a String, one token per line