Cardano.Ledger.ShelleyMA.Rules.Utxo

Documentation

data UtxoPredicateFailure era

Constructors

BadInputsUTxO !(Set (TxIn era))

OutsideValidityIntervalUTxO !ValidityInterval !SlotNo

MaxTxSizeUTxO !Integer !Integer

InputSetEmptyUTxO

FeeTooSmallUTxO !Coin !Coin

ValueNotConservedUTxO !(Delta (Value era)) !(Delta (Value era))

WrongNetwork !Network !(Set (Addr era))

WrongNetworkWithdrawal !Network !(Set (RewardAcnt era))

OutputTooSmallUTxO ![TxOut era]

UpdateFailure (PredicateFailure (PPUP era))

OutputBootAddrAttrsTooBig ![TxOut era]

TriesToForgeADA

consumed ∷ ∀ era. (ShelleyBased era, HasField "certs" (TxBody era) (StrictSeq (DCert era)), HasField "inputs" (TxBody era) (Set (TxIn era)), HasField "forge" (TxBody era) (Value era), HasField "wdrls" (TxBody era) (Wdrl era)) ⇒ PParams era → UTxO era → TxBody era → Value era

Calculate the value consumed by the transation.

This differs from the corresponding Shelley function Shelley.consumed since it also considers the "forge" field which creates or destroys non-Ada tokens.

Note that this is slightly confusing, since it also covers non-Ada assets _created_ by the transaction, depending on the sign of the quantities in the forge field.

utxoTransition ∷ ∀ era. (ShelleyBased era, STS (UTXO era), Embed (PPUP era) (UTXO era), BaseM (UTXO era) ~ ShelleyBase, Environment (UTXO era) ~ UtxoEnv era, State (UTXO era) ~ UTxOState era, Signal (UTXO era) ~ Tx era, PredicateFailure (UTXO era) ~ UtxoPredicateFailure era, HasField "certs" (TxBody era) (StrictSeq (DCert era)), HasField "inputs" (TxBody era) (Set (TxIn era)), HasField "forge" (TxBody era) (Value era), HasField "outputs" (TxBody era) (StrictSeq (TxOut era)), HasField "wdrls" (TxBody era) (Wdrl era), HasField "txfee" (TxBody era) Coin, HasField "vldt" (TxBody era) ValidityInterval, HasField "update" (TxBody era) (StrictMaybe (Update era))) ⇒ TransitionRule (UTXO era)

The UTxO transition rule for the Shelley-MA (Mary and Allegra) eras.

Orphan instances

(Crypto c, Typeable ma, ShelleyBased (ShelleyMAEra ma c), TxBody (ShelleyMAEra ma c) ~ TxBody (ShelleyMAEra ma c)) ⇒ STS (UTXO (ShelleyMAEra ma c))

Instance details

Associated Types

type State (UTXO (ShelleyMAEra ma c))

type Signal (UTXO (ShelleyMAEra ma c))

type Environment (UTXO (ShelleyMAEra ma c))

type BaseM (UTXO (ShelleyMAEra ma c)) ∷ Type → Type

type PredicateFailure (UTXO (ShelleyMAEra ma c)) = (b ∷ Type)

Methods

initialRules ∷ [InitialRule (UTXO (ShelleyMAEra ma c))]

transitionRules ∷ [TransitionRule (UTXO (ShelleyMAEra ma c))]

assertions ∷ [Assertion (UTXO (ShelleyMAEra ma c))]

renderAssertionViolation ∷ AssertionViolation (UTXO (ShelleyMAEra ma c)) → String

(Crypto c, Typeable ma) ⇒ Embed (PPUP (ShelleyMAEra ma c)) (UTXO (ShelleyMAEra ma c))

Instance details

Methods

wrapFailed ∷ PredicateFailure (PPUP (ShelleyMAEra ma c)) → PredicateFailure (UTXO (ShelleyMAEra ma c))